Add a test to ignore the Android platform.
Ignore the Android platform if the "google-services.json" is not present.

+++ scripts/after_prepare.js
@@ -33,29 +33,31 @@ if(fileExists("GoogleService-Info.plist")){
   }
 }
 
-try {
-	var contents = fs.readFileSync("google-services.json").toString();
-    fs.writeFileSync("platforms/android/google-services.json", contents);
-
-    var json = JSON.parse(contents);
-    var strings = fs.readFileSync("platforms/android/res/values/strings.xml").toString();
-
-    // strip non-default value
-    strings = strings.replace(new RegExp('<string name="google_app_id">([^\@<]+?)</string>', "i"), '')
-
-    // strip non-default value
-    strings = strings.replace(new RegExp('<string name="google_api_key">([^\@<]+?)</string>', "i"), '')
-
-    // strip empty lines
-    strings = strings.replace(new RegExp('(\r\n|\n|\r)[ \t]*(\r\n|\n|\r)', "gm"), '$1')
-
-    // replace the default value
-    strings = strings.replace(new RegExp('<string name="google_app_id">([^<]+?)</string>', "i"), '<string name="google_app_id">' + json.client[0].client_info.mobilesdk_app_id + '</string>')
-
-    // replace the default value
-    strings = strings.replace(new RegExp('<string name="google_api_key">([^<]+?)</string>', "i"), '<string name="google_api_key">' + json.client[0].api_key[0].current_key + '</string>')
-
-    fs.writeFileSync("platforms/android/res/values/strings.xml", strings);
-} catch(err) {
-    process.stdout.write(err);
+if(fileExists("google-services.json")){
+	try {
+		var contents = fs.readFileSync("google-services.json").toString();
+	    fs.writeFileSync("platforms/android/google-services.json", contents);
+	
+	    var json = JSON.parse(contents);
+	    var strings = fs.readFileSync("platforms/android/res/values/strings.xml").toString();
+	
+	    // strip non-default value
+	    strings = strings.replace(new RegExp('<string name="google_app_id">([^\@<]+?)</string>', "i"), '')
+	
+	    // strip non-default value
+	    strings = strings.replace(new RegExp('<string name="google_api_key">([^\@<]+?)</string>', "i"), '')
+	
+	    // strip empty lines
+	    strings = strings.replace(new RegExp('(\r\n|\n|\r)[ \t]*(\r\n|\n|\r)', "gm"), '$1')
+	
+	    // replace the default value
+	    strings = strings.replace(new RegExp('<string name="google_app_id">([^<]+?)</string>', "i"), '<string name="google_app_id">' + json.client[0].client_info.mobilesdk_app_id + '</string>')
+	
+	    // replace the default value
+	    strings = strings.replace(new RegExp('<string name="google_api_key">([^<]+?)</string>', "i"), '<string name="google_api_key">' + json.client[0].api_key[0].current_key + '</string>')
+	
+	    fs.writeFileSync("platforms/android/res/values/strings.xml", strings);
+	} catch(err) {
+	    process.stdout.write(err);
+	}
 }
